Deleted unnecessary files
